Abstract gateway context outside of mode package
